# docs/oir-management-api-spec-v1.yaml
openapi: 3.0.3
info:
  title: Open Image Registry - Management API
  version: 1.2.0
  description: |
    API for managing users, namespaces, repositories and access control.

    ### Security Model:
    * **Stateful JWT**: Authentication is handled via an HTTP-only cookie `auth_token`.
    * **Revocation**: The `/auth/logout` endpoint is protected and explicitly revokes the token signature.
    * **Onboarding**: Public access is granted to `/onboarding` for account activation using a setup id.

    ### Password Requirements:
    * Minimum 12 characters, maximum 64 characters
    * Must contain at least one uppercase letter
    * Must contain at least one lowercase letter
    * Must contain at least one digit
    * Must contain at least one symbol from: !@#$%^&*

    ### Username Requirements:
    * 3-32 characters
    * Allowed characters: letters, numbers, dots, underscores, hyphens
    * Pattern: ^[a-zA-Z0-9._-]{3,32}$
servers:
  - url: /api/v1
    description: Base Management API URL
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate and establish session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        '200':
          description: OK. Returns `auth_token` cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileInfo'
          headers:
            Set-Cookie:
              schema:
                { type: string, example: 'auth_token=eyJhbG...; HttpOnly; Secure; SameSite=Strict' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Terminate session and revoke token
      security:
        - cookieAuth: []
      description: |
        Requires a valid session. The middleware extracts the signature, 
        and the handler marks it as revoked in the database.
      responses:
        '200':
          description: Session revoked.
        '401': { $ref: '#/components/responses/Unauthorized' }

  # --- ONBOARDING ---
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
...
  responses:
    Unauthorized:
      description: 'Invalid, expired, or revoked session cookie.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error_code, error_message]
      properties:
        error_code:
          type: integer
          description: HTTP status code
        error_message:
          type: string
          description: Error message

    # --- USER MANAGEMENT SCHEMAS ---
...
    UserProfileInfo:
      type: object
      required:
        - user_id
        - username
        - role
      properties:
        user_id:
          type: string
        username:
          type: string
        role:
          type: string